requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitly-api==0.3
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
cssselect==1.1.0
cssutils==2.4.0
email-to==0.1.0
et-xmlfile==1.1.0
geographiclib==1.52
geopy==2.2.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
Markdown==3.3.7
msal==1.17.0
ntlm-auth==1.5.0
numpy==1.22.3
Office365-REST-Python-Client==2.3.11
openpyxl==3.0.9
pandas==1.4.1
paramiko==2.10.4
premailer==3.10.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzipcode==3.0.1
requests==2.27.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
schedule==1.1.0
SharePlum==0.5.1
sharepy==2.0.0
six==1.16.0
soupsieve==2.3.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.0.3
zipp==3.8.0
